tri training with asp ok

+++ asp_utils.py
@@ -78,8 +78,10 @@ def convert_to_consistent_answersets(preds_path, iter_number, model_number, conf
     #           ' --outf=0 -V0 --out-atomf=%s. --quiet=1,2,2'
     command = 'clingo --opt-mode=optN asp_v2/v5/p5_with_rules.lp ' + atomed_output_path + \
               ' --outf=0 -V0 --out-atomf=%s. --quiet=1,2,2'
-    shutil.rmtree(os.path.dirname(atomed_output_path.format(sent_number=0)))
-    shutil.rmtree(os.path.dirname(answerset_output_path.format(sent_number=0)))
+    if os.path.exists(os.path.dirname(atomed_output_path.format(sent_number=0))):
+        shutil.rmtree(os.path.dirname(atomed_output_path.format(sent_number=0)))
+    if os.path.exists(os.path.dirname(answerset_output_path.format(sent_number=0))):
+        shutil.rmtree(os.path.dirname(answerset_output_path.format(sent_number=0)))
     os.makedirs(os.path.dirname(atomed_output_path.format(sent_number=0)), exist_ok=True)
     os.makedirs(os.path.dirname(answerset_output_path.format(sent_number=0)), exist_ok=True)
     # Load the predictions
